Use int instead of size_t for the step parameter in nsl_geom_linesim_nthpoint().
BUG: 38197
FIXED-IN: 2.5

+++ src/backend/nsl/nsl_geom_linesim.c
@@ -275,7 +275,7 @@ double nsl_geom_linesim_douglas_peucker_variant(const double xdata[], const doub
 	return newmaxdist;
 }
 
-size_t nsl_geom_linesim_nthpoint(const size_t n, const size_t step, size_t index[]) {
+size_t nsl_geom_linesim_nthpoint(const size_t n, const int step, size_t index[]) {
 	if (step < 1) {
 		printf("step size must be > 0 (given: %zd)\n", step);
 		return 0;
